change LXC template add logging in container models

+++ container/models.py
@@ -1,7 +1,10 @@
 from django.db import models
 
 from router.models import Router
+import logging
 import lxc
+
+logger = logging.getLogger(__name__)
 
 
 class Container(models.Model):
@@ -15,18 +18,13 @@ class Container(models.Model):
         # Setup the container object
         c = lxc.Container(name)
         # Create the container rootfs
-        c.create("download", lxc.LXC_CREATE_QUIET, {"dist": "ubuntu",
-                                                    "release": "bionic",
-                                                    "arch": "amd64"})
-        # Install OpenSSH Server
-        c.attach_wait(lxc.attach_run_command,
-                      ["apt-get", "update"])
-        c.attach_wait(lxc.attach_run_command,
-                      ["apt-get", "upgrade", "-y"])
-        c.attach_wait(lxc.attach_run_command,
-                      ["apt-get", "install", "openssh-server", "-y"])
+        if not c.create("ubuntu", lxc.LXC_CREATE_QUIET):
+            logger.error("Failed to create the container rootfs")
+            raise Exception("Failed to create the container rootfs")
+        else:
+            logger.debug("Container created")
 
-        # Create container object in database
+        # Create container record in database
         container = cls(name=name, domain=domain,
                         activity=False, exposed_port=ports)
         # add rules to router
@@ -40,21 +38,36 @@ class Container(models.Model):
     def destroy(self):
         Router.objects.filter(container_name=self.name).delete()
         c = lxc.Container(self.name)
-        c.destroy()
-        self.delete()
+        if not c.destroy():
+            logger.error("Failed to destroy the container")
+            raise Exception("Failed to destroy the container")
+        else:
+            self.delete()
 
     def activate(self):
         c = lxc.Container(self.name)
-        c.start()
-        self.activity = (
-            c.state == "STARTING" or c.state == "RUNNING")
+        if not c.start():
+            logger.error("Failed to start the container")
+            raise Exception("Failed to start the container")
+        else:
+            logger.debug("Container started")
+            c.state = True
         self.save()
 
     def deactivate(self):
         c = lxc.Container(self.name)
-        c.stop()
-        self.activity = (
-            c.state == "STOPPING" or c.state == "STOPPED")
+        if not c.shutdown(30):
+            logger.error(
+                "Failed to cleanly shutdown the container, forcing...")
+            if not c.stop():
+                logger.error("Failed to kill the container")
+                raise Exception("Failed to kill the container")
+            else:
+                logger.debug("Container killed")
+                c.state = False
+        else:
+            logger.debug("Container stopped")
+            c.state = False
         self.save()
 
     def addRoute(self, port, ptype):
